Kimin work
I might not quite understand what Valtteri means of visualization (let me know if i am far away from the goal, i will try to get there??), i used on click button event to instantiate the ingredient object to the scene(because there are buttons.......with names). And if click the "Puhekupla" then it shows the ordered menu(randomly chosen,,,,,). Question: In Hampurilainen scriptableObject, what is the meaning of layers? the order of layers in unity??